Moved OTP to PwEntry
