+++ KPCLib.xunit/PassXYZLibTests.cs
@@ -68,21 +68,39 @@ namespace KPCLib.xunit
             PwGroup pg = passxyz.PxDb.RootGroup;
             foreach (var item in pg.Items)
             {
-                Debug.WriteLine($"{item.Description}\t{item.Name}");
-                if(!item.IsGroup)
+                if(item.IsGroup)
                 {
-                    if(PxDefs.IsPxEntry((PwEntry)item))
-                    {
-                        var entry = new PxEntry((PwEntry)item);
-                        if (entry.Totp == null)
-                        {
-                            entry.UpdateOtpUrl("otpauth://totp/Google%3Apxentry_test%40gmail.com?secret=JBSWY3DPEHPK3PXP&issuer=Google");
-                            var Totp = entry.Totp;
-                            Debug.WriteLine($"Token={entry.Token}");
-                            Assert.NotNull(Totp);
-                        }
-                    }
+                    Debug.WriteLine($"{item.Description}\t{item.Name}/");
                 }
+                else 
+                {
+                    Debug.WriteLine($"{item.Description}\t{item.Name}");
+                }
+            }
+        }
+
+        [Fact (Skip = "test manually")]
+        /// <summary>
+        /// This test computes a token using a secret.
+        /// The result can be verified at the below website:
+        /// https://totp.danhersam.com/
+        /// </summary>
+        public void TokenGenerateTests()
+        {
+            var entry = new PwEntry();
+            if (entry.Totp == null)
+            {
+                entry.UpdateOtpUrl("otpauth://totp/Google%3Apxentry_test%40gmail.com?secret=JBSWY3DPEHPK3PXP&issuer=Google");
+                var Totp = entry.Totp;
+                Debug.WriteLine($"Token={entry.Token}\tProgress={entry.Progress}");
+                Assert.NotNull(Totp);
+            }
+
+            int i = 0;
+            for (i = 0; i < 30; i++)
+            {
+                Debug.WriteLine($"Token={entry.Token}\tProgress={entry.Progress}");
+                System.Threading.Thread.Sleep(3000);
             }
         }
